Add per-year GAVI support column
The actual data will be added in subsequent montagu-data work; Olivia
has supplied us with the data (and OP14/OP15 include this apparently).

This is the start of a bunch of work that will totally rework the impact
calculations.

The gavi_support information is not given to modellers so no changes
should be required at the API level.